requirements.txt
# Azure Functions Core
azure-functions>=1.18.0
azure-functions-worker>=1.10.0

# Azure SDK
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-storage-blob>=12.19.0
azure-cosmos>=4.5.1

# Azure Document Intelligence (Form Recognizer)
azure-ai-formrecognizer>=3.3.0
azure-ai-documentintelligence>=1.0.0b1

# Logging and Monitoring
opencensus-ext-azure>=1.1.13
opencensus-ext-requests>=0.8.0
opencensus-ext-logging>=0.1.1

# Data Processing and Validation
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.24.0
aiohttp>=3.8.0
aiofiles>=23.0.0
pillow>=10.1.0
PyPDF2>=3.0.1

# Development and Testing
pytest>=7.4.3
pytest-asyncio>=0.23.2
pytest-mock>=3.12.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1

# Utilities
python-multipart>=0.0.6